Rename to KopyKat and write docs
