+++ ksp/src/main/kotlin/fp/serrano/kopykat/TransformativeProcessor.kt
@@ -1,4 +1,4 @@
-package fp.serrano.transformative
+package fp.serrano.kopykat
 
 import com.google.devtools.ksp.processing.CodeGenerator
 import com.google.devtools.ksp.processing.KSPLogger
@@ -7,7 +7,7 @@ import com.google.devtools.ksp.processing.SymbolProcessor
 import com.google.devtools.ksp.symbol.KSAnnotated
 import com.google.devtools.ksp.symbol.KSClassDeclaration
 import com.google.devtools.ksp.symbol.Modifier
-import fp.serrano.transformative.utils.hasGeneratedMarker
+import fp.serrano.kopykat.utils.hasGeneratedMarker
 
 internal class TransformativeProcessor(
   private val codegen: CodeGenerator,
