+++ ksp/src/main/resources/META-INF/services/com.google.devtools.ksp.processing.SymbolProcessorProvider
@@ -1,1 +1,1 @@
-fp.serrano.transformative.TransformativeProvider
+fp.serrano.kopykat.TransformativeProvider
